feat: ingredients should accept decimal and 0 nutritional values (#168)

+++ src/GlucoPilot.Api/Endpoints/Ingredients/List/ListIngredientsResponseModels.cs
@@ -14,10 +14,10 @@ public sealed record GetIngredientResponse
 {
     public required Guid Id { get; set; }
     public required string Name { get; set; }
-    public required int Carbs { get; set; }
-    public required int Protein { get; set; }
-    public required int Fat { get; set; }
-    public required int Calories { get; set; }
+    public required decimal Carbs { get; set; }
+    public required decimal Protein { get; set; }
+    public required decimal Fat { get; set; }
+    public required decimal Calories { get; set; }
     public required UnitOfMeasurement Uom { get; set; }
     public required DateTimeOffset Created { get; set; }
     public DateTimeOffset? Updated { get; set; }
